prepare for publish; fix defaults in generateClient
