+++ javascript_client/.npmignore
@@ -1,0 +1,4 @@
+node_modules/
+OperationStoreClient.js
+yarn.lock
+src/
